chore: file alocations tweaks
