+++ src/collection/domain/models/CollectionItemsQueryParams.ts
@@ -1,0 +1,7 @@
+export enum CollectionItemsQueryParams {
+  SORT = 'sort',
+  ORDER = 'order',
+  START = 'start',
+  TYPES = 'types',
+  FILTER_QUERIES = 'fqs'
+}
