--- src/collection/domain/models/GetCollectionItemsQueryParams.ts
@@ -1,19 +1,0 @@
-export enum GetCollectionItemsQueryParams {
-  SORT = 'sort',
-  ORDER = 'order',
-  START = 'start',
-  TYPES = 'types',
-  FILTER_QUERIES = 'fqs'
-}
-
-export enum SortType {
-  NAME = 'name',
-  DATE = 'date'
-}
-
-export enum OrderType {
-  ASC = 'asc',
-  DESC = 'desc'
-}
-
-export type FilterQuery = `${string}:${string}`
