+++ src/sections/collection/collection-items-panel/filter-panel/FilterPanel.tsx
@@ -5,7 +5,7 @@ import { FunnelFill } from 'react-bootstrap-icons'
 import { CollectionItemType } from '@/collection/domain/models/CollectionItemType'
 import { ItemTypeChange, TypeFilters } from './type-filters/TypeFilters'
 import { CollectionItemsFacet } from '@/collection/domain/models/CollectionItemSubset'
-import { FilterQuery } from '@/collection/domain/models/GetCollectionItemsQueryParams'
+import { FilterQuery } from '@/collection/domain/models/CollectionSearchCriteria'
 import { FacetsFilters } from './facets-filters/FacetsFilters'
 import { RemoveAddFacetFilter } from './facets-filters/FacetFilterGroup'
 import styles from './FilterPanel.module.scss'
