+++ src/sections/collection/collection-items-panel/items-list/file-card/FileCard.module.scss
@@ -47,6 +47,7 @@
     max-width: 64px;
     height: 48px;
     max-height: 48px;
+    object-fit: contain;
     vertical-align: top;
   }
